chore: rename icons paths to icons
