+++ src/app/about/contact-buttons.tsx
@@ -2,8 +2,8 @@ import Icon from '@mdi/react';
 import cx from 'classix';
 
 import { ButtonLink } from '@/components/core/link/button-link';
+import { telegram, twitterOutline } from '@/components/icons/icons';
 import { mdiEmail } from '@/components/icons/mdi';
-import { telegram, twitterOutline } from '@/components/icons/paths';
 
 const emailButtonClasses = cx(
   'hocus:bg-[rgba(211_60_48/0.08)]',
