+++ src/app/donate/sponsors-stats.tsx
@@ -1,5 +1,5 @@
+import { money } from '@/components/icons/icons';
 import { mdiHeartOutline } from '@/components/icons/mdi';
-import { money } from '@/components/icons/paths';
 import { StatCard } from '@/components/views/dashboard/stat-card/stat-card';
 
 interface SimpleSponsorsStatsProps {
