+++ src/components/views/dashboard/reactions/reactions.tsx
@@ -2,12 +2,12 @@
 
 import { useEffect, useState } from 'react';
 
+import { awardOutline } from '@/components/icons/icons';
 import {
   mdiBookmarkOutline,
   mdiHeartOutline,
   mdiThumbUpOutline,
 } from '@/components/icons/mdi';
-import { awardOutline } from '@/components/icons/paths';
 
 import { StatCard, StatCardProps } from '../stat-card/stat-card';
