+++ src/components/views/donate/buttons/buttons.tsx
@@ -2,12 +2,12 @@ import Icon from '@mdi/react';
 
 import { Heading } from '@/components/core/heading';
 import { ButtonLink } from '@/components/core/link/button-link';
+import { gift } from '@/components/icons/icons';
 import {
   mdiCreditCardChipOutline,
   mdiHeartOutline,
   mdiPizza,
 } from '@/components/icons/mdi';
-import { gift } from '@/components/icons/paths';
 
 import { AmazonLink, GitHubSponsorsLink, PayPalLink } from './buttons.styles';
