+++ src/components/views/donate/sponsor/sponsor.tsx
@@ -4,8 +4,8 @@ import type { Route } from 'next';
 import type { CSSProperties } from 'react';
 
 import { Img } from '@/components/core/img';
+import { diamond, rocket, star, unicorn } from '@/components/icons/icons';
 import { mdiCrystalBall } from '@/components/icons/mdi';
-import { diamond, rocket, star, unicorn } from '@/components/icons/paths';
 import { hexToRgb } from '@/utils/color';
 
 import {
