+++ src/components/views/projects/card/card.tsx
@@ -5,7 +5,7 @@ import cx from 'classix';
 import type { Route } from 'next';
 import { useMemo, type CSSProperties } from 'react';
 
-import { star } from '@/components/icons/paths';
+import { star } from '@/components/icons/icons';
 import { useHasMounted } from '@/hooks/use-has-mounted';
 import { useImmutableRequest } from '@/hooks/use-request';
 import { useTheme } from '@/providers/theme-provider';
